requirements.txt
django==4.2.5
mysqlclient
djangorestframework
